requirements.txt
# Install a CUDA-enabled PyTorch build separately on the server, then run:
#   pip install -r requirements.txt
#
# Example:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

PyYAML>=6.0
einops>=0.8.0
datasets>=2.19.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
scipy>=1.12.0
pyfaidx>=0.8.1.1
transformers>=4.45.0
peft>=0.12.0
wandb>=0.17.0
